Hotfix getting order from server if not found on device

+++ app/src/main/java/taxi/city/citytaxidriver/core/User.java
@@ -26,7 +26,7 @@ public class User implements Serializable {
     public String address;
     public String deviceToken;
     public Car car;
-    public double rating;
+    public float rating;
 
     private User() {}
 
@@ -57,19 +57,19 @@ public class User implements Serializable {
         String ratingSumString = json.getJSONObject("rating").getString("votes__sum");
         double ratingSum = ratingSumString == null || ratingSumString.equals("null") ? 0 : Double.valueOf(ratingSumString);
         int ratingCount = json.getJSONObject("rating").getInt("votes__count");
-        this.rating = ratingCount == 0 ? 0 : round(ratingSum/ratingCount, 1);
+        this.rating = ratingCount == 0 ? 0 : (float)round(ratingSum/ratingCount, 1);
         JSONArray car = json.getJSONArray("cars");
         if (car.length() > 0) {
             this.car = new Car(car.getJSONObject(0));
         }
     }
 
-    private double round(double value, int places) {
+    private float round(double value, int places) {
         if (places < 0) throw new IllegalArgumentException();
 
         BigDecimal bd = new BigDecimal(value);
         bd = bd.setScale(places, RoundingMode.HALF_UP);
-        return bd.doubleValue();
+        return (float)bd.doubleValue();
     }
 
     public void setRating(double a, double b) {
